pyproject.toml
[project]
name = "jarvis-planejador-financeiro"
version = "0.1.0"
description = "Assistente financeiro conversacional (Jarvis) com bot Telegram e LLM Gemini."
authors = [{ name = "Eduardo Siqueira Bonfim" }]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==20.7",
    "python-dotenv==1.0.0",
    "google-generativeai==0.3.2",
    "pandas>=2.3.3",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "apscheduler>=3.10.0",
    "google-cloud-aiplatform>=1.115.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
]


